Refactor of package structure .. Hope nothing breaks
