+++ src/main/java/dk/aau/modelardb/core/GroupBasedCompression/SegmentGenerator.java
@@ -12,9 +12,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core;
+package dk.aau.modelardb.core.GroupBasedCompression;
 
-import dk.aau.modelardb.core.models.ModelType;
+import dk.aau.modelardb.core.Models.CompressionModels.ModelType;
+import dk.aau.modelardb.core.Models.ValueDataPoint;
 import dk.aau.modelardb.core.timeseries.TimeSeries;
 import dk.aau.modelardb.core.utility.Logger;
 import dk.aau.modelardb.core.utility.ReverseBufferIterator;
