+++ src/main/java/dk/aau/modelardb/core/Models/CompressionModels/ModelType.java
@@ -12,9 +12,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core.models;
+package dk.aau.modelardb.core.Models.CompressionModels;
 
-import dk.aau.modelardb.core.ValueDataPoint;
+import dk.aau.modelardb.core.Models.ValueDataPoint;
 import dk.aau.modelardb.core.utility.Static;
 
 import java.io.Serializable;
