+++ src/main/java/dk/aau/modelardb/core/Models/CompressionModels/ModelTypeFactory.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core.models;
+package dk.aau.modelardb.core.Models.CompressionModels;
 
 import java.lang.reflect.Constructor;
 import java.util.stream.IntStream;
