+++ src/main/java/dk/aau/modelardb/core/Models/DataPoint.java
@@ -1,4 +1,4 @@
-package dk.aau.modelardb.core;
+package dk.aau.modelardb.core.Models;
 
 public abstract class DataPoint {
     private final int tid;
