+++ src/main/java/dk/aau/modelardb/core/Models/ValueDataPoint.java
@@ -12,9 +12,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core;
+package dk.aau.modelardb.core.Models;
 
-public class ValueDataPoint extends DataPoint{
+public class ValueDataPoint extends DataPoint {
 
     /**
      * Instance Variables
