+++ src/main/java/dk/aau/modelardb/core/timeseries/AsyncTimeSeries.java
@@ -14,7 +14,7 @@
  */
 package dk.aau.modelardb.core.timeseries;
 
-import dk.aau.modelardb.core.SegmentGenerator;
+import dk.aau.modelardb.core.GroupBasedCompression.SegmentGenerator;
 
 import java.io.IOException;
 import java.nio.channels.Selector;
