+++ src/main/java/dk/aau/modelardb/core/utility/ReverseBufferIterator.java
@@ -14,7 +14,7 @@
  */
 package dk.aau.modelardb.core.utility;
 
-import dk.aau.modelardb.core.ValueDataPoint;
+import dk.aau.modelardb.core.Models.ValueDataPoint;
 
 import java.util.ArrayList;
 import java.util.Iterator;
