+++ src/main/scala/dk/aau/modelardb/Main.scala
@@ -15,7 +15,7 @@
 package dk.aau.modelardb
 
 import dk.aau.modelardb.core._
-import dk.aau.modelardb.core.models.ModelTypeFactory
+import dk.aau.modelardb.core.Models.CompressionModels.ModelTypeFactory
 import dk.aau.modelardb.core.utility.{Pair, Static, ValueFunction}
 import dk.aau.modelardb.engines.{CodeGenerator, EngineFactory}
 import dk.aau.modelardb.storage.StorageFactory
