+++ src/main/scala/dk/aau/modelardb/engines/h2/H2Projector.scala
@@ -14,8 +14,9 @@
  */
 package dk.aau.modelardb.engines.h2
 
+import dk.aau.modelardb.core.GroupBasedCompression.SegmentGroup
+import dk.aau.modelardb.core.Models.ValueDataPoint
 import dk.aau.modelardb.core.utility.Static
-import dk.aau.modelardb.core.{ValueDataPoint, SegmentGroup}
 import dk.aau.modelardb.engines.{CodeGenerator, EngineUtilities, H2DataPointProjector, H2SegmentProjector}
 import org.h2.expression.aggregate.AbstractAggregate
 import org.h2.expression.condition.{Comparison, ConditionAndOr}
