+++ src/main/scala/dk/aau/modelardb/engines/h2/ViewDataPoint.scala
@@ -14,7 +14,7 @@
  */
 package dk.aau.modelardb.engines.h2
 
-import dk.aau.modelardb.core.ValueDataPoint
+import dk.aau.modelardb.core.Models.ValueDataPoint
 import org.h2.api.TableEngine
 import org.h2.command.ddl.CreateTableData
 import org.h2.command.dml.AllColumnsForPlan
