+++ src/main/scala/dk/aau/modelardb/storage/JDBCStorage.scala
@@ -14,6 +14,7 @@
  */
 package dk.aau.modelardb.storage
 
+import dk.aau.modelardb.core.GroupBasedCompression.{SegmentGroup, TimeSeriesGroup}
 import dk.aau.modelardb.core._
 import dk.aau.modelardb.core.utility.Static
 import dk.aau.modelardb.engines.h2.{H2, H2Storage}
