+++ src/test/scala/dk/aau/modelardb/core/Models/CompressionModels/AbstractModelTypeTest.scala
@@ -12,9 +12,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core.models
+package dk.aau.modelardb.core.Models.CompressionModels
 
-import dk.aau.modelardb.core.ValueDataPoint
+import dk.aau.modelardb.core.Models.ValueDataPoint
 import org.scalatest.wordspec.AnyWordSpec
 
 import java.util
