+++ src/test/scala/dk/aau/modelardb/core/Models/CompressionModels/FacebookGorillaModelTypeTest.scala
@@ -12,10 +12,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core.models
+package dk.aau.modelardb.core.Models.CompressionModels
 
-class PMC_MeanModelTypeTest extends AbstractModelTypeTest {
+class FacebookGorillaModelTypeTest extends AbstractModelTypeTest {
   override protected def getModelType(mtid: Int, errorBound: Float, lengthBound: Int): ModelType = {
-    new PMC_MeanModelType(mtid, errorBound, lengthBound)
+    new FacebookGorillaModelType(mtid, errorBound, lengthBound)
   }
 }
