+++ src/test/scala/dk/aau/modelardb/core/Models/CompressionModels/PMC_MeanModelTypeTest.scala
@@ -12,10 +12,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core.models
+package dk.aau.modelardb.core.Models.CompressionModels
 
-class FacebookGorillaModelTypeTest extends AbstractModelTypeTest {
+class PMC_MeanModelTypeTest extends AbstractModelTypeTest {
   override protected def getModelType(mtid: Int, errorBound: Float, lengthBound: Int): ModelType = {
-    new FacebookGorillaModelType(mtid, errorBound, lengthBound)
+    new PMC_MeanModelType(mtid, errorBound, lengthBound)
   }
 }
