+++ src/test/scala/dk/aau/modelardb/core/Models/CompressionModels/SwingFilterModelTypeTest.scala
@@ -12,10 +12,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dk.aau.modelardb.core.models
+package dk.aau.modelardb.core.Models.CompressionModels
 
-class UncompressedModelTypeTest extends AbstractModelTypeTest {
+class SwingFilterModelTypeTest extends AbstractModelTypeTest {
   override protected def getModelType(mtid: Int, errorBound: Float, lengthBound: Int): ModelType = {
-    new UncompressedModelType(mtid, errorBound, lengthBound)
+    new SwingFilterModelType(mtid, errorBound, lengthBound)
   }
 }
